Create mongoDBmongoose.js
This is was the first successful test of accessing, and storing information in a mongoDB using mongoose. :) Much is hard coded and will require you to install mongo, mongoose and node,js as well as configuring a new mongoDB to run. It also requires the specified mongoDB to be running.